+++ backend/src/main/java/com/kyf/mp/javaserver/modules/queuemodule/dto/AlterQueueDTO.java
@@ -9,5 +9,5 @@ import lombok.Data;
 public class AlterQueueDTO {
     @JsonProperty("queue_id")
     @NotNull(message = "缺少 queue_id")
-    private Integer queueId;
+    private Long queueId;
 }
